Move relevant files to Durable\Tasks folder
