+++ src/Durable/Commands/StartDurableTimerCommand.cs
@@ -10,6 +10,7 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Durable.Commands
     using System;
     using System.Collections;
     using System.Management.Automation;
+    using Microsoft.Azure.Functions.PowerShellWorker.Durable.Tasks;
 
     /// <summary>
     /// Start the Durable Functions timer
