+++ src/Durable/StartDurableExternalEventListener.cs
@@ -9,6 +9,7 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Durable
 {
     using System.Collections;
     using System.Management.Automation;
+    using Microsoft.Azure.Functions.PowerShellWorker.Durable.Tasks;
 
     /// <summary>
     /// Start the Durable External Event Listener
