+++ src/Durable/Tasks/ActivityInvocationTask.cs
@@ -5,7 +5,7 @@
 
 #pragma warning disable 1591 // "Missing XML comments for publicly visible type"
 
-namespace Microsoft.Azure.Functions.PowerShellWorker.Durable
+namespace Microsoft.Azure.Functions.PowerShellWorker.Durable.Tasks
 {
     using System;
     using System.Linq;
@@ -13,6 +13,8 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Durable
 
     using WebJobs.Script.Grpc.Messages;
 
+    using Microsoft.Azure.Functions.PowerShellWorker;
+    using Microsoft.Azure.Functions.PowerShellWorker.Durable;
     using Microsoft.Azure.Functions.PowerShellWorker.Durable.Actions;
 
     public class ActivityInvocationTask : DurableTask
